[refactor]ディレクトリ構成と命名を整理
