+++ app/components/timetables/columns/base_component.html.erb
